Add card functionality and API methods
