+++ src/components/Main.js
@@ -1,7 +1,6 @@
 import React from "react";
-import { api } from "../utils/Api";
 import Card from "./Card";
-import { useEffect, useState, useContext } from "react";
+import { useContext } from "react";
 import CurrentUserContext from "./contexts/CurrentUserContext";
 
 function Main({
@@ -9,34 +8,11 @@ function Main({
   onEditProfileClick,
   onAddPlaceClick,
   onCardClick,
+  cards,
+  onCardLike,
+  onCardDelete,
 }) {
-  const [cards, setCards] = useState([]);
   const { currentUser } = useContext(CurrentUserContext);
-
-  function handleCardLike(card) {
-    const isLiked = card.likes.some((i) => i._id === currentUser._id);
-    api.likeCard(card._id, isLiked).then((res) => {
-      setCards((state) => state.map((c) => (c._id === card._id ? res : c)));
-    });
-  }
-
-  function handleCardDelete(card) {
-    api.deleteCard(card._id).then((res) => {
-      setCards((state) => state.filter((c) => c._id !== card._id));
-    });
-  }
-
-  const fetchInitialCards = () => {
-    api.getInitialCards().then((res) => {
-      if (Array.isArray(res)) {
-        setCards(res);
-      }
-    });
-  };
-
-  useEffect(() => {
-    fetchInitialCards();
-  }, []);
 
   return (
     <main className="container">
@@ -73,8 +49,8 @@ function Main({
             key={card._id}
             card={card}
             onCardClick={onCardClick}
-            onCardLike={handleCardLike}
-            onCardDelete={handleCardDelete}
+            onCardLike={onCardLike}
+            onCardDelete={onCardDelete}
           />
         ))}
       </section>
